Remove ChannelStateLogic interface
